+++ src/Command/FetchCensusCommand.php
@@ -6,8 +6,7 @@ use App\Entity\Midata\CensusGroup;
 use App\Model\CommandStatistics;
 use App\Repository\Midata\CensusGroupRepository;
 use App\Repository\Midata\GroupTypeRepository;
-use App\Service\CensusAPIService;
-use App\Service\GroupStructureAPIService;
+use App\Service\Census\CensusAPIService;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
